Adopt experimental MaterialFadeThrough to enter/exit motion spec
