+++ library/src/main/java/soup/compose/material/motion/experimental/MotionSpec.kt
@@ -19,12 +19,16 @@ import androidx.compose.animation.EnterExitState
 import androidx.compose.animation.EnterTransition
 import androidx.compose.animation.ExitTransition
 import androidx.compose.animation.ExperimentalAnimationApi
+import androidx.compose.animation.core.FiniteAnimationSpec
 import androidx.compose.animation.core.Transition
+import androidx.compose.animation.core.animateFloat
+import androidx.compose.animation.core.spring
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.graphicsLayer
 
 @ExperimentalAnimationApi
 class EnterMotionSpec(
@@ -50,3 +54,47 @@ class TransitionExtra<E>(
         )
     }
 }
+
+@ExperimentalAnimationApi
+fun scaleIn(
+    initialScale: Float,
+    animationSpec: FiniteAnimationSpec<Float>,
+): TransitionExtra<Float> = scaleTransitionExtra(
+    initialScale = initialScale,
+    animationSpec = animationSpec
+)
+
+@ExperimentalAnimationApi
+fun scaleOut(
+    targetScale: Float,
+    animationSpec: FiniteAnimationSpec<Float>,
+): TransitionExtra<Float> = scaleTransitionExtra(
+    initialScale = targetScale,
+    animationSpec = animationSpec
+)
+
+@ExperimentalAnimationApi
+private fun scaleTransitionExtra(
+    initialScale: Float,
+    targetScale: Float = 1f,
+    animationSpec: FiniteAnimationSpec<Float> = spring(),
+): TransitionExtra<Float> = TransitionExtra(
+    animateExtra = {
+        animateFloat(
+            transitionSpec = { animationSpec },
+            label = "scale"
+        ) {
+            when (it) {
+                EnterExitState.PreEnter -> initialScale
+                EnterExitState.Visible -> targetScale
+                EnterExitState.PostExit -> initialScale
+            }
+        }
+    },
+    modifierByExtra = { scale ->
+        Modifier.graphicsLayer {
+            scaleX = scale
+            scaleY = scale
+        }
+    }
+)
